yolov5_fastapi/requirements.txt
fastapi
uvicorn
numpy
opencv-python
onnxruntime
pillow
